Move PWM and LiDAR to modules folder
